feat: 重构面试功能并优化云存储URL处理
refactor(面试服务): 重构面试开始接口，使用DTO接收参数并支持职位类型ID
fix(用户服务): 修复用户信息接口的token刷新逻辑
style(小程序): 清理未使用的面试配置文件和样式
perf(云存储): 优化云存储URL到HTTPS格式的转换逻辑
docs: 移除未使用的AI API配置和注释

+++ ai-resume-service/src/main/java/com/aicv/airesume/repository/JobTypeRepository.java
@@ -10,7 +10,7 @@ import java.util.Optional;
  * 职位类型数据访问接口
  */
 @Repository
-public interface JobTypeRepository extends JpaRepository<JobType, Long> {
+public interface JobTypeRepository extends JpaRepository<JobType, Integer> {
 
     /**
      * 根据职位名称查询职位类型
@@ -25,4 +25,5 @@ public interface JobTypeRepository extends JpaRepository<JobType, Long> {
      * @return 职位类型列表
      */
     java.util.List<JobType> findByDomainId(Long domainId);
+    
 }
